Converted doubles to points

+++ src/main/java/processing/style/MLA9.java
@@ -45,24 +45,24 @@ public class MLA9 extends StyleGuide {
      *  MLA uses an indentation padding of 0.5 inches
      */
     @Override
-    protected double indentationPadding() {
-        return 0.5d;
+    protected float indentationPadding() {
+        return 0.5f;
     }
 
     /**
      *  MLA uses a margin of 1 inch
      */
     @Override
-    public double margin() {
-        return 1;
+    public float margin() {
+        return 1f;
     }
 
     /**
      *  MLA uses double-spacing
      */
     @Override
-    public double spacing() {
-        return 2d;
+    public float spacing() {
+        return 2f;
     }
 
     /**
@@ -83,8 +83,8 @@ public class MLA9 extends StyleGuide {
     }
 
     @Override
-    public double numerationMargin() {
-        return 0.5d;
+    public float numerationMargin() {
+        return 0.5f;
     }
 
     @Override
@@ -99,8 +99,8 @@ public class MLA9 extends StyleGuide {
 
     // todo: check
     @Override
-    public double paragraphIndentation() {
-        return 0.5d;
+    public float paragraphIndentation() {
+        return 0.5f;
     }
 
     @Override
